chore: add i18next to storybook
